Update to v0.1.1 with Edition 2024, dependency cleanup, and comprehensive documentation updates
- Upgrade Cargo.toml edition to "2024" (stable since Rust 1.85.0) and version to 0.1.1
- Remove unused dependencies (crossterm, rayon, rdev, serde_json), retain as comments for future phases
- Keep serde dependency (required by ocr_llm.rs for API serialization)
- Add tokio, reqwest, base64, dialoguer dependencies for LLM OCR and async runtime
- Delete obsolete docs: cli_site_option.md, engine_api_fix

+++ src/engine.rs
@@ -29,7 +29,7 @@ pub fn analyze_position(fen: &str) -> Result<(String, String)> {
     // Step 4: Extract evaluation score (PSQT after best move; white-positive)
     let mut eval_board = board.shallow_clone();
     eval_board.apply_move(best_move);
-    let raw_eval = eval_board.psq().mg() as i32;
+    let raw_eval = eval_board.psq().mg();
 
     // Negate for side-to-move perspective (if Black to move, flip sign)
     let eval_score = if board.turn() == tanton::Player::Black { -raw_eval } else { raw_eval };
